fix(supabase): match_documents rompía siempre por mismatch de tipo REAL/double precision
El operador <=> de pgvector devuelve double precision; "1 - (...)" hereda
ese tipo, pero la función declara similarity REAL. Sin cast explícito,
Postgres rechaza la función con "structure of query does not match
function result type" en cada llamada — confirmado en vivo contra
hsglmdarztrshihmzfph (proyecto ya con el schema creado). Esto rompía
toda búsqueda RAG de la Bóveda (vault-process "search" y el contexto
del orchestrator). Cast a ::REAL en el SELECT y en el WHERE.

+++ supabase/migrations/001_initial_schema.sql
@@ -225,9 +225,9 @@ BEGIN
     dc.id,
     dc.document_id,
     dc.content,
-    1 - (dc.embedding <=> query_embedding) AS similarity
+    (1 - (dc.embedding <=> query_embedding))::REAL AS similarity
   FROM doc_chunks dc
-  WHERE 1 - (dc.embedding <=> query_embedding) > similarity_threshold
+  WHERE (1 - (dc.embedding <=> query_embedding))::REAL > similarity_threshold
   ORDER BY dc.embedding <=> query_embedding
   LIMIT match_count;
 END;
